changefeedccl: move metrics into their own file
No change, only code movement.

Release note: None
